* Allow overriding the detected CPU vendor to allow more self tests

* Add plugin private flags

This matches what we can do with FuDevice.


* rustgen: bound the array index in generated getters

* build(deps): bump codecov/codecov-action from 6 to 7

Bumps [codecov/codecov-action](https://github.com/codecov/codecov-action) from 6 to 7.
- [Release notes](https://github.com/codecov/codecov-action/releases)
- [Changelog](https://github.com/codecov/codecov-action/blob/main/CHANGELOG.md)
- [Commits](https://github.com/codecov/codecov-action/compare/v6...v7)

---
updated-dependencies:
- dependency-name: codecov/codecov-action
  dependency-version: '7'
  dependency-type: direct:production
  update-type: version-update:semver-major
...


* build(deps): bump github/codeql-action from 4.36.1 to 4.36.2

Bumps [github/codeql-action](https://github.com/github/codeql-action) from 4.36.1 to 4.36.2.
- [Release notes](https://github.com/github/codeql-action/releases)
- [Changelog](https://github.com/github/codeql-action/blob/main/CHANGELOG.md)
- [Commits](https://github.com/github/codeql-action/compare/87557b9c84dde89fdd9b10e88954ac2f4248e463...8aad20d150bbac5944a9f9d289da16a4b0d87c1e)

---
updated-dependencies:
- dependency-name: github/codeql-action
  dependency-version: 4.36.2
  dependency-type: direct:production
  update-type: version-update:semver-patch
...


* trivial: Allow fu_config_set_value_internal() to remove keys

The `value` parameter is marked as nullable and so this should be possible.

* trivial: Add some context helpers for future code

* dell-kestrel: Fix the config value default in the man file

* test: Fix the config value defaults in the man file

* trivial: Make the man config defaults more resilient

* Simplify the engine by only loading the config object once

We were loading `FuConfig` via `FuContext` and then loading `FuEngineConfig`
(which itself derives from `FuConfig`) again from `FuEngine` -- but the
underlying `GKeyFile` dictionaries were actually different objects.

This made debugging very confusing as setting an override on one `FuConfig`
didn't have the expected outcome -- and meant that we had to be careful about
the context lifecycle in the tests.

Merge them together and use `FuConfig` directly from `FuEngine`.

* trivial: Remove an unused FuQuirksLoadFlags

* trivial: Add NoCache as a context flag

This is used also used at runtime and not just load-time.

* trivial: Rename FuContextHwidFlags to FuContextLoadFlags

* trivial: Add a FuContextLoadFlags argument to fu_context_load_quirks()

* trivial: Make loading the context simpler

In various places we're loading the context quirks, config, and usually both.

Merge `fu_context_load_hwinfo()` and `fu_context_load_quirks()` into a single
`fu_context_load()` that can do both.

This reduces the 'you're holding it wrong' problem when writing tests.

* trivial: Allow setting IgnoreEfivarsFreeSpace at runtime

Also, abstract out some common code.

* trivial: Use FuContext when creating a FuRemoteList

* trivial: Use FuContextLoadFlags when loading the remotes

* trivial: Squash fu_context_load_path_store() into fu_context_load()

* Partially revert "Use fu_memstrsafe() in more places for safety"

This partially reverts commit e9a9a006ce53bd130a66a613362b8eb5632a9c10 as it
caused regressions with empty msgpack strings, and UTF-8 msgpack strings.

* msr: Add unit tests for the plugin

Use emulated FuDeviceEvent pread data to control which MSR registers
are supported, avoiding any direct struct access in the self tests.


* pci-bcr: Add unit tests for the plugin


* elan-ts: Add new plugin for Elan touchscreen


* upower: Add unit tests for the plugin


* trivial: make glib get included with `-isystem`

This way we prevent some tools such as clang-tidy from complaining about
the contents of glib headers, which we can't fix.

* trivial: update clang-tidy config

Use more readable array format for `Checks:`, disable some more checks
for now and treat all warnings as errors so it's more useful for CI.

* trivial: add clang-tidy CI workflow

* Install the db updates on more hardware

HP have told me that any machine with the 2023 CA in dbDefault is new enough to
not get bricked when deploying the db update at runtime, so build a new
instance ID for those vendors that includes the default certificate hashes.

This requires a different component on the LVFS for machines that use this,
but that's probably a good idea from an analytics and demotion point of view.

Also, export FuX509Certificate setters and add a write vfunc that generates
self-signed DER certificates using GnuTLS, allowing the uefi-db self-tests to
generate test certificates at runtime.


* csv: bound the column index in fu_csv_entry_get_value_by_column_id

* linux-lockdown: Never add NULL as the LinuxLockdown report metadata value

* upower: Do not run the unit tests if dbus-daemon is not available

* trivial: pin all github actions to commit hashes

Should make supply chain attacks of GH actions a little more difficult.

* Log unexpected netlink recvfrom socket read errors

When running without systemd-udevd (e.g, Android), if the raw netlink socket
encounters unexpected read failures, explicitly log them rather than failing
silently.

* Expand netlink socket buffer to prevent packet loss during event floods

When running without systemd-udevd under heavy kernel hotplug activity (such as
composite USB gadget initialization), uevents can flood the raw netlink socket.

This expands SO_RCVBUF to 8MB to prevent packet dropping (ENOBUFS).

* trivial: actions/checkout no persist-credentials

This is apparently an insecure default for this action, and we
presumably never need it.

* trivial: Fix clang-tidy 'redundant cast to the same type'

* trivial: Fix clang-tidy 'statement should have braces'

* trivial: Fix clang-tidy 'redundant parentheses around expression'

* trivial: Fix clang-tidy 'variable can be made static'

* trivial: Fix one -Wlogical-op warning

* Fix clang-tidy 'repeated branch body in conditional chain'

* Fix clang-tidy 'function is called without explicitly comparing result'

* trivial: Ignore clang-tidy bugprone-unchecked-string-to-number-conversion

* Fix clang-tidy 'signed char conversion'

* trivial: Translated using Weblate (Turkish)

Currently translated at 48.7% (435 of 892 strings)

Translate-URL: https://hosted.weblate.org/projects/fwupd/fwupd/tr/
Translation: fwupd/fwupd

* trivial: Fix clang-tidy 'readability-use-concise-preprocessor-directives'

* trivial: Fix clang-tidy 'readability-redundant-control-flow'

* trivial: Disable invalid clang-tidy exclusion

* trivial: Fix clang-tidy 'readability-else-after-return'

* trivial: Fix clang-tidy 'readability-inconsistent-declaration-parameter-name'

* trivial: Disable invalid clang-tidy exclusion

* trivial: Fix clang-tidy 'readability-math-missing-parentheses'

* trivial: Fix clang-tidy 'readability-isolate-declaration'

* trivial: Disable invalid clang-tidy exclusion

* trivial: Fix clang-tidy 'readability-avoid-const-params-in-decls'

* trivial: Disable invalid clang-tidy exclusion

* trivial: Fix clang-tidy 'bugprone-too-small-loop-variable'

* trivial: Disable invalid clang-tidy exclusion

* trivial: Fix some clang-tidy 'bugprone-macro-parentheses'

* trivial: Disable invalid clang-tidy exclusion

---------

+++ libfwupdplugin/fu-coswid-firmware.c
@@ -127,7 +127,7 @@ fu_coswid_firmware_parse_meta(FuCborItem *item, gpointer user_data, GError **err
 
 		fu_cbor_item_map_index(item, i, &item_key, &item_value);
 		if (!fu_coswid_read_tag(item_key, &tag_id, error)) {
-			g_prefix_error(error, "failed to parse meta tag %u: ", (guint)i);
+			g_prefix_error(error, "failed to parse meta tag %u: ", i);
 			return FALSE;
 		}
 		if (tag_id == FU_COSWID_TAG_SUMMARY) {
